sumcheck module
